+++ autogen/templates/message.hpp
@@ -6,7 +6,6 @@
 #include <fmt/format.h>
 
 #include <chrono>
-#include <span>
 #include <string_view>
 
 #include "roq/compat.hpp"
@@ -15,6 +14,7 @@
 #include "roq/message_info.hpp"
 #include "roq/name.hpp"
 #include "roq/numbers.hpp"
+#include "roq/span.hpp"
 #include "roq/string_types.hpp"
 #include "roq/trace.hpp"