+++ include/roq/execution_instruction.hpp
@@ -6,6 +6,8 @@
 
 #include <fmt/format.h>
 
+#include <cassert>
+
 namespace roq {
 
 //! Enumeration of execution types
@@ -29,6 +31,7 @@ struct fmt::formatter<roq::ExecutionInstruction> {
   template <typename Context>
   auto format(const roq::ExecutionInstruction &value, Context &context) {
     using namespace std::literals;
+#if __cplusplus >= 202002L
     std::string_view name{[&]() {
       switch (value) {
         using enum roq::ExecutionInstruction;
@@ -47,6 +50,25 @@ struct fmt::formatter<roq::ExecutionInstruction> {
       }
       return "<UNKNOWN>"sv;
     }()};
+#else
+    std::string_view name{[&]() {
+      switch (value) {
+        case roq::ExecutionInstruction::UNDEFINED:
+          return "UNDEFINED"sv;
+        case roq::ExecutionInstruction::PARTICIPATE_DO_NOT_INITIATE:
+          return "PARTICIPATE_DO_NOT_INITIATE"sv;
+        case roq::ExecutionInstruction::CANCEL_IF_NOT_BEST:
+          return "CANCEL_IF_NOT_BEST"sv;
+        case roq::ExecutionInstruction::DO_NOT_INCREASE:
+          return "DO_NOT_INCREASE"sv;
+        case roq::ExecutionInstruction::DO_NOT_REDUCE:
+          return "DO_NOT_REDUCE"sv;
+        default:
+          assert(false);
+      }
+      return "<UNKNOWN>"sv;
+    }()};
+#endif
     return fmt::format_to(context.out(), "{}"sv, name);
   }
 };