+++ include/roq/modify_order.hpp
@@ -8,7 +8,6 @@
 #include <fmt/format.h>
 
 #include <chrono>
-#include <span>
 #include <string_view>
 
 #include "roq/compat.hpp"
@@ -17,6 +16,7 @@
 #include "roq/message_info.hpp"
 #include "roq/name.hpp"
 #include "roq/numbers.hpp"
+#include "roq/span.hpp"
 #include "roq/string_types.hpp"
 #include "roq/trace.hpp"